ex résolu

+++ TP7/pyramide/pyramide.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int max(int a, int b){
+unsigned long long int max(unsigned long long int a, unsigned long long int b){
     if (a>b)
         return a;
     else
@@ -10,14 +10,14 @@ int max(int a, int b){
 
 int main(){
     int n;
-    unsigned long int res=0;
+    unsigned long long int res=0;
     scanf("%i", &n);
     if (n<1 || n>30){
         printf("Nombre pas dans le bon intervalle !\n");
         return 42;
     }
     int dim = n;
-    unsigned int tab_init[dim][dim];
+    unsigned long long int tab_init[dim][dim];
     for (int i=0; i<dim; i++){
         for (int j=0; j<dim; j++){
                 tab_init[i][j]=0;
@@ -26,18 +26,19 @@ int main(){
     for (int i=0; i<dim; i++){
         for (int j=0; j<=i; j++){
             //printf("i,j = %i %i\n", i, j);
-            scanf("%i", &tab_init[i][j]);
+            scanf("%llu", &tab_init[i][j]);
         }
     }
+
     /*printf("\n");
     for (int k=0; k<dim; k++){
         for (int l=0; l<dim; l++){
-                printf("%i ", tab_init[k][l]);
+                printf("%llu ", tab_init[k][l]);
         }
         printf("\n");
     }*/
     
-    unsigned int tab[dim][dim];
+    unsigned long long int tab[dim][dim];
     for (int i=0; i<dim; i++){
         for (int j=0; j<dim; j++){
                 tab[i][j]=0;
@@ -57,12 +58,12 @@ int main(){
     /*printf("\n");
     for (int k=0; k<dim; k++){
         for (int l=0; l<dim; l++){
-                printf("%i ", tab[k][l]);
+                printf("%llu ", tab[k][l]);
         }
         printf("\n");
     }*/
 
     res=tab[0][0];
-    printf("\n%li\n", res);
+    printf("\n%llu\n", res);
     return 0;
 }
